+++ packages/wonder-blocks-banner/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @khanacademy/wonder-blocks-banner
 
+## 4.1.24
+
+### Patch Changes
+
+- Updated dependencies [52393f6]
+    - @khanacademy/wonder-blocks-button@10.2.1
+
+## 4.1.23
+
+### Patch Changes
+
+- Updated dependencies [e9498ec]
+- Updated dependencies [cca3595]
+    - @khanacademy/wonder-blocks-icon-button@10.3.0
+    - @khanacademy/wonder-blocks-button@10.2.0
+
+## 4.1.22
+
+### Patch Changes
+
+- Updated dependencies [8468d8d]
+- Updated dependencies [308696e]
+- Updated dependencies [dc34e0c]
+- Updated dependencies [7c59ffd]
+- Updated dependencies [2cace08]
+- Updated dependencies [f714517]
+- Updated dependencies [8468d8d]
+    - @khanacademy/wonder-blocks-tokens@10.5.0
+    - @khanacademy/wonder-blocks-icon@5.2.0
+    - @khanacademy/wonder-blocks-button@10.1.0
+    - @khanacademy/wonder-blocks-icon-button@10.2.2
+    - @khanacademy/wonder-blocks-link@9.1.6
+    - @khanacademy/wonder-blocks-typography@4.2.1
+
 ## 4.1.21
 
 ### Patch Changes