+++ packages/wonder-blocks-popover/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @khanacademy/wonder-blocks-popover
 
+## 6.1.8
+
+### Patch Changes
+
+- Updated dependencies [e9498ec]
+    - @khanacademy/wonder-blocks-icon-button@10.3.0
+    - @khanacademy/wonder-blocks-modal@7.1.22
+    - @khanacademy/wonder-blocks-tooltip@4.1.22
+
+## 6.1.7
+
+### Patch Changes
+
+- Updated dependencies [8468d8d]
+    - @khanacademy/wonder-blocks-tokens@10.5.0
+    - @khanacademy/wonder-blocks-icon-button@10.2.2
+    - @khanacademy/wonder-blocks-modal@7.1.21
+    - @khanacademy/wonder-blocks-styles@0.2.13
+    - @khanacademy/wonder-blocks-tooltip@4.1.21
+    - @khanacademy/wonder-blocks-typography@4.2.1
+
 ## 6.1.6
 
 ### Patch Changes